centerOnVision command; started panToZeroCommand

+++ src/main/java/frc/robot/RobotMap.java
@@ -37,9 +37,9 @@ public class RobotMap {
   //shooter constants
   /* TODO: get a more accurate value for this */
   public final static int shooterPanMotorEncoderTicksPerRotation = 3977;
-  public final static int shooterXResolution = 640;
-  public final static int shooterYResolution = 480; 
-  public final static int shooterResolutionAcceptableError = 20;
+  public final static double shooterXResolution = 640;
+  public final static double shooterYResolution = 480; 
+  public final static double shooterResolutionAcceptableError = 20;
 
   // Control panel constants
   // TODO: Get actual motor ID
